Added 1D Pong sketch
The 1D Pong game is still in progress. This initial sketch bounces the
light back and forth between two sides of the Adafruit RGB LED strip,
increasing in speed before maxing out and staying at a certain level
(5ms delays). The color of the light changes each time it hits a wall,
and the user can attach a button to try and hit it at the right time.